+++ src/main/java/io/nzbee/entity/bag/item/domain/IBagItemDomainDTOMapper.java
@@ -6,7 +6,7 @@ public interface IBagItemDomainDTOMapper<T, E, D> {
 
 	D DTOToDo(Bag bag, T dto);
 
-	D DTOToDo(Bag bag, T dto, int quantity);
+	D DTOToDo(Bag bag, T dto, Long quantity);
 	
 	E doToEntity(D d);
